# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hope Apartments API",
    "version": "1.0.0",
    "description": "Swagger documentation for the Hope Apartments API. Partner apartment access uses X-API-Key. Internal docs are protected with admin session or Bearer JWT auth."
  },
  "servers": [
    {
      "url": "https://api.hope-apartments.de",
      "description": "Railway development"
    }
  ],
...
  "paths": {
...
    "/openapi.json": {
      "get": {
        "tags": ["System"],
        "summary": "OpenAPI specification",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OpenAPI document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/docs": {
      "get": {
        "tags": ["System"],
        "summary": "Swagger UI",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Interactive API documentation",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Current authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Current user" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/apartments": {
      "get": {
        "tags": ["Apartments"],
        "summary": "Fetch live apartment data",
        "description": "Returns normalized apartments from onOffice. Partner access requires a valid X-API-Key.",
        "security": [
          {
            "partnerApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Partner API key"
          }
        ],
        "responses": {
          "200": {
            "description": "Normalized apartment payload",
            "headers": {
              "x-data-source": {
                "schema": {
                  "type": "string"
                },
                "description": "Data source marker",
                "example": "live-onoffice"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                }
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                }
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apartments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Apartment"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": ["apartments", "meta"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Another sync is already running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Upstream fetch or mapping failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys": {
      "get": {
        "tags": ["API Keys"],
        "summary": "List API keys",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "API key list" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "post": {
        "tags": ["API Keys"],
        "summary": "Create API key",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": { "description": "API key created" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api-keys/{id}": {
      "get": {
        "tags": ["API Keys"],
        "summary": "Read API key by id",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "API key details" },
          "404": { "description": "Not found" }
        }
      },
      "patch": {
        "tags": ["API Keys"],
        "summary": "Update API key",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "API key updated" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api-keys/{id}/revoke": {
      "post": {
        "tags": ["API Keys"],
        "summary": "Revoke API key",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "API key revoked" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api-keys/{id}/reactivate": {
      "post": {
        "tags": ["API Keys"],
        "summary": "Reactivate API key",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "API key reactivated" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api-keys/{id}/rotate": {
      "post": {
        "tags": ["API Keys"],
        "summary": "Rotate API key",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "API key rotated" },
          "404": { "description": "Not found" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "partnerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Apartment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "roomsTotal": {
            "type": "number",
            "nullable": true,
            "example": 3
          },
          "bedrooms": {
            "type": "number",
            "nullable": true,
            "example": 2
          },
          "bathrooms": {
            "type": "number",
            "nullable": true,
            "example": 1
          },
          "areaSqft": {
            "type": "number",
            "nullable": true,
            "example": 538.19
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Photo"
            }
          },
          "features": {
            "$ref": "#/components/schemas/Features"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "locationDescription": {
            "type": "string",
            "nullable": true
          },
          "equipmentDescription": {
            "type": "string",
            "nullable": true
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "rent": {
            "$ref": "#/components/schemas/Rent"
          },
          "deposit": {
            "type": "string",
            "nullable": true
          },
          "floorLevel": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "address",
          "roomsTotal",
          "bedrooms",
          "bathrooms",
          "areaSqft",
          "photos",
          "features",
          "description",
          "locationDescription",
          "equipmentDescription",
          "availability",
          "rent",
          "deposit",
          "floorLevel"
        ]
      },
      "Address": {
        "type": "object",
        "properties": {
          "buildingNumber": {
            "type": "string",
            "nullable": true
          },
          "streetName": {
            "type": "string",
            "nullable": true
          },
          "neighborhood": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["buildingNumber", "streetName", "neighborhood", "city", "zipCode"]
      },
      "Photo": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "originalname": {
            "type": "string",
            "nullable": true
          },
          "modified": {
            "nullable": true,
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          }
        },
        "required": ["url", "type", "title", "originalname", "modified"]
      },
      "Features": {
        "type": "object",
        "properties": {
          "elevator": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "balcony": {
            "type": "boolean"
          },
          "furnished": {
            "type": "boolean",
            "nullable": true
          }
        },
        "required": ["elevator", "balcony", "furnished"]
      },
      "Availability": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "nullable": true,
            "example": "2026-05-01"
          },
          "until": {
            "type": "string",
            "nullable": true,
            "example": "2026-12-31"
          }
        },
        "required": ["from", "until"]
      },
      "Rent": {
        "type": "object",
        "properties": {
          "warmRent": {
            "type": "number",
            "nullable": true,
            "example": 1200
          },
          "coldRent": {
            "type": "number",
            "nullable": true,
            "example": 1000
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          }
        },
        "required": ["warmRent", "coldRent", "currency"]
      },
      "ResponseMeta": {
        "type": "object",
        "properties": {
          "requestedBy": {
            "type": "string",
            "example": "partner-a"
          },
          "count": {
            "type": "integer",
            "example": 84
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "durationMs": {
            "type": "integer",
            "example": 3000
          }
        },
        "required": ["requestedBy", "count", "startedAt", "finishedAt", "durationMs"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["error", "message"]
      }
    }
  }
}